Get tests in order
